Refactor the Neighbors classes
