+++ src/cellmapper/__init__.py
@@ -2,8 +2,9 @@ from importlib.metadata import version
 
 from .logging import logger
 from .model.cellmapper import CellMapper
+from .model.kernel import Kernel
 from .model.neighbors import Neighbors
 
-__all__ = ["logger", "CellMapper", "Neighbors"]
+__all__ = ["logger", "CellMapper", "Kernel", "Neighbors"]
 
 __version__ = version("cellmapper")
