+++ src/cellmapper/model/neighbors.py
@@ -1,256 +1,422 @@
+from dataclasses import dataclass
+from functools import cached_property
 from typing import Literal
 
 import numpy as np
-from scipy.sparse import csr_matrix
+from scipy.sparse import coo_matrix, csr_matrix
+from umap.umap_ import fuzzy_simplicial_set
 
 from cellmapper.constants import PackageConstants
 from cellmapper.logging import logger
-from cellmapper.model._knn_backend import get_backend
-from cellmapper.model.neighbors_results import NeighborsResults
-from cellmapper.utils import extract_neighbors_from_distances
 
 
+@dataclass
 class Neighbors:
-    """Class to compute and store nearest neighbors."""
+    """Nearest neighbors results data store.
 
-    def __init__(self, xrep: np.ndarray, yrep: np.ndarray | None = None, is_self_mapping: bool | None = None):
+    Adapted from the scib-metrics package: https://github.com/YosefLab/scib-metrics.
+    Extended to support non-square matrices and potentially varying number of neighbors per cell.
+    This class stores the results of nearest neighbor searches and provides methods to compute
+    adjacency matrices and connectivities using various kernels.
+
+    Features:
+    ---------
+    - Multiple connectivity kernels (Gaussian, adaptive Gaussian, inverse distance, etc.)
+    - Support for both square (self-mapping) and non-square (cross-mapping) matrices
+    - Configurable self-edge inclusion for square matrices in connectivity computations
+    - Symmetrization options for creating undirected graphs
+    - Robust handling of variable neighbor counts and invalid entries
+
+
+    Attributes
+    ----------
+    distances
+        Array of distances to the nearest neighbors, excluding self-edges.
+        For square matrices, self-edges are automatically removed during initialization
+        to ensure consistent storage format across different k-NN algorithms.
+    indices
+        Array of indices of the nearest neighbors, excluding self-edges.
+        For square matrices, self-edges are automatically removed during initialization
+        to ensure consistent storage format across different k-NN algorithms.
+    n_targets
+        Number of target samples. If None, it is assumed to be the same as
+        the number of samples in the indices array.
+
+    Notes
+    -----
+    The `n_neighbors` property always refers to non-self neighbors (self-edges are automatically
+    removed during initialization for consistent storage). When self-edges are included via
+    connectivity methods with `self_edges=True`, the resulting arrays will have shape
+    (n_samples, n_neighbors + 1), but the `n_neighbors` property remains unchanged.
+    """
+
+    distances: np.ndarray
+    indices: np.ndarray
+    n_targets: int | None = None
+
+    def __post_init__(self):
         """
-        Initialize the Neighbors class.
+        Post-initialization logic for NeighborsResults.
+
+        Ensures that `n_targets` is set correctly, validates array shapes,
+        and removes self-edges from square matrices for consistent storage.
+        """
+        if self.indices.shape != self.distances.shape:
+            raise ValueError("Indices and distances must have the same shape.")
+
+        if self.n_targets is None:
+            # Assume square adjacency matrix if `n_targets` is not provided
+            self.n_targets = self.indices.shape[0]
+
+        # Remove self-edges for consistent storage (square matrices only)
+        if self.is_square:
+            self._remove_self_edges_from_storage()
+
+    @property
+    def n_samples(self) -> int:
+        """Number of samples (cells)."""
+        return self.indices.shape[0]
+
+    @property
+    def n_neighbors(self) -> int:
+        """Number of neighbors."""
+        return self.indices.shape[1]
+
+    @property
+    def shape(self) -> tuple[int, int]:
+        """Shape of the adjacency/graph matrices (n_samples, n_targets)."""
+        return (self.n_samples, self.n_targets or self.n_samples)
+
+    @property
+    def is_square(self) -> bool:
+        """Whether this represents a square matrix (self-mapping)."""
+        return self.n_samples == (self.n_targets or self.n_samples)
+
+    def _create_sparse_matrix_from_arrays(
+        self, values: np.ndarray, indices: np.ndarray, valid_mask: np.ndarray, dtype=np.float64
+    ) -> csr_matrix:
+        """
+        Helper method to create a sparse matrix from values and indices with filtering of invalid entries.
 
         Parameters
         ----------
-        xrep
-            Representation of the reference dataset.
-        yrep
-            Representation of the query dataset. If None, self-mapping will be used.
-        is_self_mapping
-            Explicitly specify if this is a self-mapping case. If None, will be inferred
-            from whether yrep is None.
-        """
-        self.xrep = xrep
-        # Use xrep for self-mapping if yrep is None
-        self.yrep = yrep if yrep is not None else xrep
-
-        # Initialize neighbor result containers
-        self.xx: NeighborsResults | None = None
-        self.yy: NeighborsResults | None = None
-        self.xy: NeighborsResults | None = None
-        self.yx: NeighborsResults | None = None
-
-        # Flag to track if this is a self-mapping case
-        # Use explicit parameter if provided, otherwise infer from yrep
-        if is_self_mapping is not None:
-            self._is_self_mapping = is_self_mapping
-        else:
-            self._is_self_mapping = yrep is None
-
-    @classmethod
-    def from_distances(cls, distances_matrix: csr_matrix, remove_last_neighbor: bool = False) -> "Neighbors":
-        """
-        Create a Neighbors object from a pre-computed distances matrix.
-
-        Parameters
-        ----------
-        distances_matrix
-            Sparse distance matrix, typically from adata.obsp['distances']
-        remove_last_neighbor
-            If True, removes the last neighbor from the distances matrix.
-            This is useful for direct comparisons with scanpy, which uses a
-            different convention for neighbor counts.
+        values
+            Values for the sparse matrix, same shape as indices.
+        indices
+            Indices array (may include self-edges).
+        valid_mask
+            Boolean mask of valid entries, same shape as values and indices.
+        dtype
+            Data type for the matrix values.
 
         Returns
         -------
-        Neighbors
-            A new Neighbors object with pre-computed neighbor information.
-            Self-edge handling is performed automatically by NeighborsResults during initialization.
+        csr_matrix
+            Sparse matrix with only valid entries.
         """
-        assert distances_matrix is not None, "distances_matrix must be provided"
-        # Extract indices and distances from the sparse matrix
-        indices, distances = extract_neighbors_from_distances(distances_matrix)
+        # Flatten all arrays
+        flat_indices = indices.ravel()
+        flat_values = values.ravel()
+        valid_entries = valid_mask.ravel()
 
-        if remove_last_neighbor:
-            # Remove the last neighbor (last column) from indices and distances
-            indices = indices[:, :-1]
-            distances = distances[:, :-1]
+        # Extract valid data
+        valid_indices = flat_indices[valid_entries]
+        valid_values = flat_values[valid_entries]
 
-            logger.info("Removed last neighbor from distances matrix for compatibility with scanpy conventions.")
+        # Create row indices
+        rows = np.repeat(np.arange(self.n_samples), indices.shape[1])
+        rows = rows[valid_entries]
 
-        # Create a minimal Neighbors object for self-mapping
-        n_cells = distances_matrix.shape[0]  # type: ignore
-        placeholder_rep = np.zeros((n_cells, 1))
-        neighbors = cls(xrep=placeholder_rep)
+        # Create CSR matrix with only valid entries
+        return csr_matrix((valid_values.astype(dtype), (rows, valid_indices)), shape=self.shape)
 
-        # Create a NeighborsResults object with the extracted data
-        neighbors_result = NeighborsResults(distances=distances, indices=indices)
+    @cached_property
+    def knn_graph_distances(self, dtype=np.float64) -> csr_matrix:
+        """
+        Return the sparse weighted adjacency matrix of distances.
 
-        # For self-mapping, all neighbor objects should be the same
-        neighbors.xx = neighbors_result
-        neighbors.yy = neighbors_result
-        neighbors.xy = neighbors_result
-        neighbors.yx = neighbors_result
+        Parameters
+        ----------
+        dtype
+            Data type for the matrix values.
 
-        # Mark as self-mapping
-        neighbors._is_self_mapping = True
+        Returns
+        -------
+        csr_matrix
+            Sparse matrix of distances (shape: n_samples x n_targets).
+        """
+        # Get distances, indices, and valid mask without self-edges by default
+        distances, indices, valid_mask = self._get_distances_and_indices(self_edges=False)
 
-        logger.info("Created Neighbors object from distances matrix with %d cells", n_cells)
+        # Create sparse matrix with distances as values
+        return self._create_sparse_matrix_from_arrays(distances, indices, valid_mask, dtype=dtype)
 
-        return neighbors
-
-    def compute_neighbors(
+    def knn_graph_connectivities(
         self,
-        n_neighbors: int = 30,
-        method: Literal["sklearn", "pynndescent", "rapids", "faiss"] = "sklearn",
-        metric: str = "euclidean",
-        random_state: int = 0,
-        only_yx: bool = False,
+        kernel: Literal["gauss", "scarches", "random", "inverse_distance", "equal", "umap"] = "gauss",
+        self_edges: bool = False,
+        dtype=np.float64,
         **kwargs,
-    ):
+    ) -> csr_matrix:
         """
-        Compute nearest neighbors using either sklearn or rapids.
+        Compute connectivities using the specified kernel.
 
         Parameters
         ----------
-         n_neighbors
-            Number of nearest neighbors.
-        method
-            Method to use for computing neighbors.
-        metric
-            Distance metric to use for nearest neighbors.
-        random_state
-            Random state for reproducibility.
-        only_yx
-            If True, only compute the xy neighbors. In self-mapping mode, this is
-            automatically set to True for efficiency since all neighbor matrices
-            contain the same information.
+        kernel
+            Connectivity kernel to use. Supported: 'gauss', 'scarches', 'random', 'inverse_distance', 'equal', 'umap'.
+        self_edges
+            Control self-edges in the neighbor graph for square matrices:
+            - True: Include self-edges in the connectivity computation
+            - False: Exclude self-edges from the connectivity computation
+        dtype
+            Data type for the matrix values.
         **kwargs
-            Additional keyword arguments to pass to the underlying k-NN algorithm.
-            These are method-specific and will be passed directly to the algorithm's
-            constructor or fitting method.
-
-            For pynndescent, scanpy-style defaults are applied:
-            - n_jobs: -1 (use all CPU cores)
-            - n_trees: min(64, 5 + round(n_samples^0.5 / 20.0)) (per dataset)
-            - n_iters: max(5, round(log2(n_samples))) (per dataset)
+            Additional keyword arguments for kernel computation.
 
         Returns
         -------
-        None
-
-        Notes
-        -----
-        Updates the following attributes:
-
-        - ``xx``: Nearest neighbors results for reference to reference.
-        - ``yy``: Nearest neighbors results for query to query.
-        - ``xy``: Nearest neighbors results for reference to query.
-        - ``yx``: Nearest neighbors results for query to reference.
-
-        In self-mapping mode, all four matrices will reference the same NeighborsResults
-        object for memory efficiency.
-
-        Examples
-        --------
-        >>> neighbors = Neighbors(xrep, yrep)
-        >>> # sklearn with custom parameters
-        >>> neighbors.compute_neighbors(method="sklearn", algorithm="ball_tree", leaf_size=20)
-        >>> # pynndescent with custom parameters (inherits scanpy-style defaults)
-        >>> neighbors.compute_neighbors(method="pynndescent", n_trees=32, verbose=True)
+        csr_matrix
+            Sparse matrix of connectivities (shape: n_samples x n_targets).
         """
-        # Optimize for self-mapping: only compute yx and reuse for all matrices
-        if self._is_self_mapping:
-            only_yx = True
-            logger.info(
-                "Self-mapping mode detected. Computing only yx neighbors for efficiency "
-                "(all neighbor matrices will contain the same information)."
-            )
+        # Check if self-mapping only kernel is used for non-square matrices
+        if kernel in PackageConstants.SELF_MAPPING_ONLY_KERNELS and not self.is_square:
+            raise ValueError(f"Kernel '{kernel}' is only supported for self-mapping (square matrices)")
 
-        # issue a warning if using sklearn with large datasets
-        if method == "sklearn" and (
-            self.xrep.shape[0] > PackageConstants.SKLEARN_WARNING_CUTOFF
-            or self.yrep.shape[0] > PackageConstants.SKLEARN_WARNING_CUTOFF
-        ):
-            logger.warning(
-                "Using sklearn for neighbor search with large dataset (%d cells). "
-                "Consider using approximate k-NN search (e.g. pynndescent) or GPU acceleration (e.g. faiss or rapids)",
-                self.xrep.shape[0],
-            )
+        # Compute connectivities using the specified kernel (all kernels now return sparse matrices)
+        if kernel == "umap":
+            conn_matrix = self._compute_umap_kernel(self_edges, **kwargs)
+        else:
+            conn_matrix = self._compute_kernel_values(kernel, self_edges, dtype=dtype, **kwargs)
 
-        # use strategy pattern to reduce duplication
-        logger.info("Using %s to compute %d neighbors.", method, n_neighbors)
-        backend_x = get_backend(method, n_neighbors=n_neighbors, metric=metric, random_state=random_state, **kwargs)
-        backend_x.fit(self.xrep)
+        return conn_matrix
 
-        if only_yx:
-            dists, idx = backend_x.query(self.yrep, k=n_neighbors)
-            self.yx = NeighborsResults(distances=dists, indices=idx, n_targets=self.xrep.shape[0])
-            return
-
-        backend_y = get_backend(method, n_neighbors=n_neighbors, metric=metric, random_state=random_state, **kwargs)
-        backend_y.fit(self.yrep)
-
-        x_d, x_i = backend_x.query(self.xrep, k=n_neighbors)
-        y_d, y_i = backend_y.query(self.yrep, k=n_neighbors)
-        xy_d, xy_i = backend_y.query(self.xrep, k=n_neighbors)
-        yx_d, yx_i = backend_x.query(self.yrep, k=n_neighbors)
-
-        self.xx = NeighborsResults(distances=x_d, indices=x_i, n_targets=None)
-        self.yy = NeighborsResults(distances=y_d, indices=y_i, n_targets=None)
-        self.xy = NeighborsResults(distances=xy_d, indices=xy_i, n_targets=self.yrep.shape[0])
-        self.yx = NeighborsResults(distances=yx_d, indices=yx_i, n_targets=self.xrep.shape[0])
-
-    def get_adjacency_matrices(
-        self, symmetrize: bool = False, self_edges: bool = True
-    ) -> tuple[csr_matrix, csr_matrix, csr_matrix, csr_matrix]:
+    def _compute_kernel_values(
+        self,
+        kernel: Literal["gauss", "scarches", "random", "inverse_distance", "equal"],
+        self_edges: bool,
+        dtype=np.float64,
+        **kwargs,
+    ) -> csr_matrix:
         """
-        Compute unweighted adjacency matrices for all k-NN graphs.
+        Helper method to compute kernel values based on distances.
 
         Parameters
         ----------
-        symmetrize
-            If True, make self-terms (xx, yy) symmetrize. Cross-terms (xy, yx) are not affected.
+        kernel
+            Kernel type to use for computing connectivities.
         self_edges
-            Control self-edges for self-terms (xx, yy). Cross-terms (xy, yx) are not affected.
-            - True: Include self-edges (set diagonal entries to 1)
-            - False: Exclude self-edges (set diagonal entries to 0)
-            - None: Leave as-is (preserve original neighbor graph structure)
+            Whether to include self-edges in the computation.
+        dtype
+            Data type for the matrix values.
+        **kwargs
+            Additional arguments for kernel computation.
+
+        Returns
+        -------
+        csr_matrix
+            Sparse matrix of connectivity values.
+        """
+        # Get distances, indices, and valid mask with appropriate self-edge handling
+        distances, indices, valid_mask = self._get_distances_and_indices(self_edges)
+
+        # Initialize empty connectivities array
+        connectivities = np.zeros_like(distances)
+
+        # Extract finite distances for parameter calculation
+        finite_distances = distances[valid_mask]
+        if len(finite_distances) == 0:
+            raise ValueError("No finite distances found in the neighborhood graph")
+
+        if kernel == "gauss":
+            # Calculate sigma using only finite distances
+            sigma = np.mean(finite_distances)
+            # Apply Gaussian kernel to valid entries
+            connectivities[valid_mask] = np.exp(-(finite_distances**2) / (2 * sigma**2))
+
+        elif kernel == "equal":
+            # Set connectivities to 1 for valid entries
+            connectivities[valid_mask] = 1.0
+
+        elif kernel == "scarches":
+            # Calculate sigma using only finite distances
+            sigma = np.std(finite_distances)
+            sigma = (2.0 / sigma) ** 2
+            # Apply scArches kernel to valid entries
+            connectivities[valid_mask] = np.exp(-finite_distances / sigma)
+
+        elif kernel == "random":  # this is for testing purposes
+            # Generate random connectivities for valid entries
+            connectivities[valid_mask] = np.random.rand(np.sum(valid_mask))
+
+        elif kernel == "inverse_distance":
+            # Get epsilon parameter with default
+            epsilon = kwargs.get("epsilon", 1e-8)
+            # Apply inverse distance kernel to valid entries
+            connectivities[valid_mask] = 1.0 / (finite_distances + epsilon)
+
+        else:
+            raise ValueError(
+                f"Unknown kernel: {kernel}. Supported kernels are: 'gauss', 'scarches', 'random', 'inverse_distance', 'equal'."
+            )
+
+        # Create and return sparse matrix
+        return self._create_sparse_matrix_from_arrays(connectivities, indices, valid_mask, dtype=dtype)
+
+    def _compute_umap_kernel(self, self_edges: bool, **kwargs) -> csr_matrix:
+        """
+        Compute UMAP fuzzy simplicial set connectivities following scanpy implementation.
+
+        This calls umap-learn's fuzzy_simplicial_set function with a dummy data matrix
+        to compute the connectivity weights based on the k-NN graph structure.
+
+        Parameters
+        ----------
+        self_edges
+            Whether to include self-edges in the computation.
+        **kwargs
+            Additional parameters for UMAP kernel:
+            - set_op_mix_ratio: float, default 1.0
+            - local_connectivity: float, default 1.0
+
+        Returns
+        -------
+        csr_matrix
+            Sparse connectivity matrix.
+        """
+        # Get distances, indices, and valid mask with appropriate self-edge handling
+        distances, indices, _valid_mask = self._get_distances_and_indices(self_edges)
+
+        # UMAP kernel requires true k-NN graphs (no padding with -1)
+        if np.any(indices == -1):
+            raise ValueError("UMAP kernel requires true k-NN graphs (all cells must have exactly k neighbors)")
+
+        # Extract UMAP-specific parameters
+        set_op_mix_ratio = kwargs.get("set_op_mix_ratio", 1.0)
+        local_connectivity = kwargs.get("local_connectivity", 1.0)
+
+        # Create dummy data matrix (scanpy approach)
+        X = coo_matrix((self.n_samples, 1))
+
+        # Call UMAP's fuzzy_simplicial_set
+        result = fuzzy_simplicial_set(
+            X,
+            indices.shape[1],
+            None,  # random_state
+            None,  # metric
+            knn_indices=indices,
+            knn_dists=distances,
+            set_op_mix_ratio=set_op_mix_ratio,
+            local_connectivity=local_connectivity,
+        )
+
+        # Extract the connectivity matrix (first element of tuple)
+        connectivities_sparse = result[0]
+
+        # Return as CSR matrix
+        return connectivities_sparse.tocsr()
+
+    def boolean_adjacency(self, dtype=np.float64, self_edges: bool = False) -> csr_matrix:
+        """
+        Construct a boolean adjacency matrix from neighbor indices.
+
+        Parameters
+        ----------
+        dtype
+            Data type for the matrix values.
+        self_edges
+            Control self-edges in the neighbor graph for square matrices:
+            - True: Include self-edges in the adjacency computation
+            - False: Exclude self-edges from the adjacency computation
+
+        Returns
+        -------
+        csr_matrix
+            Boolean adjacency matrix (shape: n_samples x n_targets), with 1 for each neighbor relationship.
+        """
+        # Get distances, indices, and valid mask with appropriate self-edge handling
+        _distances, indices, valid_mask = self._get_distances_and_indices(self_edges)
+
+        # Create array of ones with same shape as indices
+        ones = np.ones_like(indices, dtype=dtype)
+
+        # Create sparse matrix with ones as values for valid entries
+        adj_matrix = self._create_sparse_matrix_from_arrays(ones, indices, valid_mask, dtype=dtype)
+
+        return adj_matrix
+
+    def _remove_self_edges_from_storage(self):
+        """Remove self-edges from stored distances and indices if present.
+
+        Only applies to square matrices. Checks if the first column contains
+        self-references (0, 1, 2, 3, ...) and removes it if found.
+        """
+        if not self.is_square:
+            return  # Nothing to do for non-square matrices
+
+        # Check if first column contains self-references (0, 1, 2, 3, ...)
+        expected_self_indices = np.arange(self.n_samples)
+        has_self_edges = np.array_equal(self.indices[:, 0], expected_self_indices)
+
+        if has_self_edges:
+            # Remove first column (self-edges) and keep remaining columns
+            self.indices = self.indices[:, 1:]
+            self.distances = self.distances[:, 1:]
+
+    def _get_distances_and_indices(self, self_edges: bool = False) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+        """Get distances and indices with optional self-edge handling, plus valid entries mask.
+
+        Parameters
+        ----------
+        self_edges
+            Control self-edge inclusion:
+            - True: Add self-edges as first column with distance 0.0
+            - False: Return stored arrays (without self-edges)
 
         Returns
         -------
         tuple
-            Unweighted adjacency matrices (xx, yy, xy, yx).
-
-        Notes
-        -----
-        The symmetrize and self_edges parameters only apply to self-terms (xx, yy) since
-        these represent within-dataset neighborhoods. Cross-terms (xy, yx) represent
-        between-dataset relationships where symmetry and self-edges are not meaningful.
+            Modified distances, indices, and valid_mask arrays based on self_edges parameter.
+            When self_edges=True, arrays will have shape (n_samples, n_neighbors + 1).
+            Valid mask identifies entries that are neither -1 indices nor infinite distances.
         """
-        if self._is_self_mapping:
-            if self.yx is None:
-                raise ValueError("Neighbors must be computed before accessing adjacency matrices.")
-            yx_adj = self.yx.boolean_adjacency(self_edges=self_edges, symmetrize=symmetrize)
-            xx_adj = yx_adj
-            yy_adj = yx_adj
-            xy_adj = yx_adj
+        if not self.is_square and self_edges:
+            logger.warning(
+                "self_edges=True is only applicable for square matrices (self-mapping). Ignoring self_edges parameter."
+            )
+
+        if self.is_square and self_edges:
+            distances, indices = self._add_self_edges(self.distances, self.indices)
         else:
-            if self.xx is None or self.yy is None or self.xy is None or self.yx is None:
-                raise ValueError("Neighbors must be computed before accessing adjacency matrices.")
+            distances, indices = self.distances, self.indices
 
-            # self-terms (within-dataset neighborhoods)
-            xx_adj = self.xx.boolean_adjacency(self_edges=self_edges, symmetrize=symmetrize)
-            yy_adj = self.yy.boolean_adjacency(self_edges=self_edges, symmetrize=symmetrize)
+        # Create valid entries mask
+        valid_mask = (indices != -1) & np.isfinite(distances)
 
-            # Cross-terms (between-dataset neighborhoods)
-            xy_adj = self.xy.boolean_adjacency(self_edges=False, symmetrize=False)
-            yx_adj = self.yx.boolean_adjacency(self_edges=False, symmetrize=False)
+        return distances, indices, valid_mask
 
-        return xx_adj, yy_adj, xy_adj, yx_adj
+    def _add_self_edges(self, distances: np.ndarray, indices: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
+        """Add self-edges as first column with distance 0.0.
 
-    def __repr__(self):
-        """Return a string representation of the Neighbors object."""
-        return (
-            f"Neighbors(xrep_shape={self.xrep.shape}, yrep_shape={self.yrep.shape}, "
-            f"xx={self.xx is not None}, yy={self.yy is not None}, "
-            f"xy={self.xy is not None}, yx={self.yx is not None}, "
-            f"self_mapping={self._is_self_mapping})"
-        )
+        Parameters
+        ----------
+        distances
+            Distance array without self-edges
+        indices
+            Indices array without self-edges
+
+        Returns
+        -------
+        tuple
+            Arrays with self-edges prepended as first column.
+            Shape increases from (n_samples, n_neighbors) to (n_samples, n_neighbors + 1).
+        """
+        # Create self-edge columns
+        self_indices = np.arange(self.n_samples, dtype=indices.dtype).reshape(-1, 1)
+        self_distances = np.zeros((self.n_samples, 1), dtype=distances.dtype)
+
+        # Prepend self-edges to existing arrays (no truncation)
+        new_indices = np.column_stack([self_indices, indices])
+        new_distances = np.column_stack([self_distances, distances])
+
+        return new_distances, new_indices
