+++ tests/model/test_neighbors_results.py
@@ -2,13 +2,13 @@ import numpy as np
 import pytest
 from scipy.sparse import csr_matrix
 
-from cellmapper.model.neighbors_results import NeighborsResults
+from cellmapper.model.neighbors import Neighbors
 
 
 class TestNeighborsResults:
     def test_neighborsresults_init_shape(self, sample_distances, sample_indices):
         # Should not raise
-        nr = NeighborsResults(distances=sample_distances, indices=sample_indices)
+        nr = Neighbors(distances=sample_distances, indices=sample_indices)
         assert nr.n_samples == 3
         # After initialization, self-edges are removed for square matrices, so n_neighbors = 1
         assert nr.n_neighbors == 1
@@ -18,10 +18,10 @@ class TestNeighborsResults:
         # indices shape mismatch should raise
         bad_indices = np.array([[0, 1, 2], [1, 2, 0], [2, 0, 1]])
         with pytest.raises(ValueError):
-            NeighborsResults(distances=sample_distances, indices=bad_indices)
+            Neighbors(distances=sample_distances, indices=bad_indices)
 
     def test_knn_graph_distances(self, sample_distances, sample_indices):
-        nr = NeighborsResults(distances=sample_distances, indices=sample_indices)
+        nr = Neighbors(distances=sample_distances, indices=sample_indices)
         mat = nr.knn_graph_distances
         assert isinstance(mat, csr_matrix)
         assert mat.shape == (3, 3)
@@ -30,7 +30,7 @@ class TestNeighborsResults:
 
     @pytest.mark.parametrize("kernel", ["gauss", "scarches", "random", "inverse_distance"])
     def test_knn_graph_connectivities_kernels(self, sample_distances, sample_indices, kernel):
-        nr = NeighborsResults(distances=sample_distances, indices=sample_indices)
+        nr = Neighbors(distances=sample_distances, indices=sample_indices)
         mat = nr.knn_graph_connectivities(kernel=kernel)
         assert isinstance(mat, csr_matrix)
         assert mat.shape == (3, 3)
@@ -39,7 +39,7 @@ class TestNeighborsResults:
             assert np.all(mat.data > 0)
 
     def test_boolean_adjacency(self, sample_distances, sample_indices):
-        nr = NeighborsResults(distances=sample_distances, indices=sample_indices)
+        nr = Neighbors(distances=sample_distances, indices=sample_indices)
         mat = nr.boolean_adjacency()
         assert isinstance(mat, csr_matrix)
         assert mat.shape == (3, 3)
